added AvailableService
