+++ src/app/controllers/AvailabeController.js
@@ -1,15 +1,4 @@
-import { Op } from 'sequelize';
-import {
-  startOfDay,
-  endOfDay,
-  setHours,
-  setMinutes,
-  setSeconds,
-  format,
-  isAfter,
-} from 'date-fns';
-
-import Appointment from '../models/Appointment';
+import AvailableService from '../services/AvailableService';
 
 class AvailableController {
   async index(req, resp) {
@@ -21,47 +10,9 @@ class AvailableController {
 
     const searchDate = Number(date);
 
-    const appointments = await Appointment.findAll({
-      where: {
-        provider_id: req.params.providerId,
-        canceled_at: null,
-        date: {
-          [Op.between]: [startOfDay(searchDate), endOfDay(searchDate)],
-        },
-      },
-    });
-
-    // TODO - Jogar para tabela no banco
-    const schedule = [
-      '08:00',
-      '09:00',
-      '10:00',
-      '11:00',
-      '12:00',
-      '13:00',
-      '14:00',
-      '15:00',
-      '16:00',
-      '17:00',
-      '18:00',
-      '19:00',
-      '20:00',
-    ];
-
-    const available = schedule.map(time => {
-      const [hour, minute] = time.split(':');
-      const value = setSeconds(
-        setMinutes(setHours(searchDate, hour), minute),
-        0
-      );
-
-      return {
-        time,
-        value: format(value, "yyyy-MM-dd'T'HH:mm:ssxxx"),
-        available:
-          isAfter(value, new Date()) &&
-          !appointments.find(ap => format(ap.date, 'HH:mm') === time),
-      };
+    const available = await AvailableService.run({
+      provider_id: req.params.providerId,
+      date: searchDate,
     });
 
     return resp.json(available);
